refactor(core): split analyze.rs into analyze.rs + analyze_focused.rs (#1293)
