+++ crates/aptu-coder-core/src/lib.rs
@@ -30,6 +30,8 @@
 #![cfg_attr(test, allow(clippy::expect_used))]
 
 pub mod analyze;
+/// Internal implementation details for focused analysis. Use the `analyze` module for the public API.
+pub(crate) mod analyze_focused;
 pub mod cache;
 pub mod cache_disk;
 pub mod completion;
